% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{add_empty_rows}
\alias{add_empty_rows}
\title{Adds rows to data.table for timestamps with missing data}
\usage{
add_empty_rows(raw, vars = c("b", "g"), first_timestamp = NULL,
  last_timestamp = NULL)
}
\arguments{
\item{raw}{data-input (data.table or data.frame)}

\item{vars}{the variables with missing data}

\item{first_timestamp}{start date}

\item{last_timestamp}{end date}
}
\value{
data.table rows for each timestamps
}
\description{
Adds rows to data.table for timestamps with missing data,
Removes impossible speed-values and rows with double timestamps
}

