% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{add_time_col}
\alias{add_time_col}
\title{Adds additional columns for avg-imputations}
\usage{
add_time_col(data)
}
\arguments{
\item{data}{data-input}
}
\value{
data.table
}
\description{
Adds additional columns for avg-imputations: dow, time, isholiday and isbridge
}

