% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{aggregate_to_interval}
\alias{aggregate_to_interval}
\title{Aggregate data to other time-intervals}
\usage{
aggregate_to_interval(data, vars = c("b", "g"), interval = 5)
}
\arguments{
\item{data}{the input-datatable (including sensor, z, posix and variables with missing data)}

\item{vars}{the variables with missing data}

\item{interval}{the interval of the target-datatable}
}
\value{
a datatable with bigger time-intervals
}
\description{
Take first elements of every column (sensor, z, posix)
and use mean for variables with missing data
}

