% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_builder.R
\name{create_avg_week}
\alias{create_avg_week}
\title{Create average week}
\usage{
create_avg_week(data, vars = c("b", "g"), interval = 5)
}
\arguments{
\item{data}{the input data.table (containing sensor, z, b, g)}

\item{vars}{the variable names to be averaged}

\item{interval}{the data interval}
}
\value{
data.table with avg, rollmean10, rollmean20 for every variable in var
}
\description{
Create average week (dow, time) excluding holidays and bridgedays
and create average holidays and bridgedays
}

