% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_gaps}
\alias{get_gaps}
\title{Finds gaps in a vector and returns their startpoints and lengths}
\usage{
get_gaps(x)
}
\arguments{
\item{x}{the vector}
}
\value{
a data.table with startpoints and lengths of all gaps in x
}
\description{
Finds gaps in a vector and returns their startpoints and lengths
}

