% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_helper.R
\name{get_holidays}
\alias{get_holidays}
\title{All Holidays (Dresden), including 24.12.}
\usage{
get_holidays(year = 2014)
}
\arguments{
\item{year}{the year you want to get holidays from (only 2014 available)}
}
\value{
vector of characters
}
\description{
All Holidays (Dresden), including 24.12.
}

