% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_choice.R
\name{get_std_choices}
\alias{get_std_choices}
\title{Get the recommendet strategies depending on variable and gap}
\usage{
get_std_choices()
}
\value{
list of lists of strategies (one list per variable)
}
\description{
Each choice gets a maximal gapsize (limit), the strategy as function,
required parameters of the stratgy and its name
}

