% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_choice.R
\name{get_test_choices}
\alias{get_test_choices}
\title{Get test strategies to be tested with every gapsize}
\usage{
get_test_choices()
}
\value{
list of lists of strategies (one list per variable)
}
\description{
Each choice gets a maximal gapsize (limit), the strategy as function,
required parameters of the stratgy and its name
}

