% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute}
\alias{impute}
\title{organizes the whole imputation-process}
\usage{
impute(data, vars = c("b", "g"), choices = get_std_choices())
}
\arguments{
\item{data}{the data.table incl. variables with missing data}

\item{vars}{vector of names of variables with missing data}

\item{choices}{list of strategies with limits and parameters}
}
\value{
data.table with sensor-id, timestamp and imputed variables
}
\description{
gets the incomplete dataset, imputes it according to strategy-choices and returns complete dataset
}

