% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_strategies.R
\name{impute_arima}
\alias{impute_arima}
\title{Impute missing values of one variable by a arima-model with kalman-filter}
\usage{
impute_arima(data, var, p = 2, d = 1, q = 2, ...)
}
\arguments{
\item{data}{the data.table incl. variables with missing data}

\item{var}{the name of the variable with missing data}

\item{p}{the order of AR}

\item{d}{the order of I}

\item{q}{the order of MA}

\item{...}{not used}
}
\value{
vector with complete data of variable
}
\description{
Impute missing values of one variable by a arima-model with kalman-filter
}

