% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_strategies.R
\name{impute_avg}
\alias{impute_avg}
\title{Impute missing values of one variable by substitution with avg-week or holi- and bridgeday}
\usage{
impute_avg(data, var, roll_interval = 1, ...)
}
\arguments{
\item{data}{the data.table incl. variables with missing data}

\item{var}{the name of the variable with missing data}

\item{roll_interval}{which aggregation-interval of avg-week (5, 10 or 20)}

\item{...}{not used}
}
\value{
vector with complete data of variable
}
\description{
Impute missing values of one variable by substitution with avg-week or holi- and bridgeday
}

