% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_combined}
\alias{impute_combined}
\title{imputes the missing data of one variable}
\usage{
impute_combined(data, var, choices)
}
\arguments{
\item{data}{the data.table incl. variables with missing data}

\item{var}{name of variable with missing data}

\item{choices}{list of strategies with limits and parameters for this variable}
}
\value{
vector with complete data of variable
}
\description{
imputes the missing data of one variable
}

