% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{impute_first_value}
\alias{impute_first_value}
\title{Impute first value if it is NA}
\usage{
impute_first_value(data, vars)
}
\arguments{
\item{data}{the data.table where avg-cols should be added}

\item{vars}{the variables with missing data}
}
\value{
the data.table without NA at the beginning
}
\description{
Imputes the first value of each variable if it is NA.
This is because some methods can't handle NA at the beginning.
}

