% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_strategies.R
\name{impute_kalman}
\alias{impute_kalman}
\title{Impute missing values of one variable by na.kalman-function (imputeTS-package)}
\usage{
impute_kalman(data, var, ...)
}
\arguments{
\item{data}{the data.table incl. variables with missing data}

\item{var}{the name of the variable with missing data}

\item{...}{not used}
}
\value{
vector with complete data of variable
}
\description{
Impute missing values of one variable by na.kalman-function (imputeTS-package)
}

