% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_var}
\alias{impute_var}
\title{applies the chosen imputation-function}
\usage{
impute_var(FUN, para)
}
\arguments{
\item{FUN}{function to be applied}

\item{para}{parameters for the chosen function (incl. variable-name, data.table with all data and additional parameters)}
}
\value{
completed data as vector
}
\description{
applies the chosen imputation-function
}

