% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_strategies.R
\name{impute_weighted_ma}
\alias{impute_weighted_ma}
\title{Impute missing values of one variable by Weighted Moving Average (imputeTS-package)}
\usage{
impute_weighted_ma(data, var, k = 10, weighting = "simple", ...)
}
\arguments{
\item{data}{the data.table incl. variables with missing data}

\item{var}{the name of the variable with missing data}

\item{k}{size of moving window (Expands to both sides of the center element e.g. k=2 means 4 observations (2 left, 2 right))}

\item{weighting}{Weighting to be used. Accepts the following input:
"simple" - Simple Moving Average (SMA)
"linear" - Linear Weighted Moving Average (LWMA)
"exponential" - Exponential Weighted Moving Average (EWMA)
(explanation from imputeTS-package)}

\item{...}{not used}
}
\value{
vector with complete data of variable
}
\description{
Impute missing values of one variable by Weighted Moving Average (imputeTS-package)
}

