% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_helper.R
\name{is_bridge}
\alias{is_bridge}
\title{Check date for bridge-day}
\usage{
is_bridge(x)
}
\arguments{
\item{x}{data.table with dow and isholiday}
}
\value{
bool-vector with T = is bridge-day and F = is no bridge-day
}
\description{
Check date for bridge-day (day between holiday and weekend) (in Dresden)
}

