% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_helper.R
\name{is_holiday}
\alias{is_holiday}
\title{Check date for holiday}
\usage{
is_holiday(x, holidays = get_holidays())
}
\arguments{
\item{x}{character-vecor with dates to be checked}

\item{holidays}{character-vecor with dates of holidays}
}
\value{
bool-vector with T = is holiday and F = is no holiday
}
\description{
Tests if date is a holiday (in Dresden)
}

