% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocessing of raw data}
\usage{
preprocess_data(file, vars = c("b", "g"), header = F, sep = ",", ...)
}
\arguments{
\item{file}{the csv-file to load}

\item{vars}{the variables to be imputed}

\item{header}{logical, if csv-input-file has header}

\item{sep}{the separator between columns in the csv-input}

\item{...}{can be first_timestamp and last_timestamp}
}
\value{
data.table with all timestamps, averages and incomplete variables
}
\description{
read from csv, add rows, aggregate, create and add averages
}

