% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{read_data_csv}
\alias{read_data_csv}
\title{Read csv to data.table}
\usage{
read_data_csv(file, sep = ",", header = F, select = c(1, 2, 3, 4),
  vars = c("b", "g"))
}
\arguments{
\item{file}{filename to read from}

\item{sep}{separator within rows in csv-file}

\item{header}{does csv-file contain header}

\item{select}{which columns to read}

\item{vars}{the variables with missing data}
}
\value{
filled data.table
}
\description{
Read csv to data.table
}

