% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{remove_long_zero}
\alias{remove_long_zero}
\title{Exclude long 0-sequences}
\usage{
remove_long_zero(data, var = "b", limit = 120)
}
\arguments{
\item{data}{the table}

\item{var}{the variable to be checked}

\item{limit}{the max. allowed 0-sequence-length}
}
\value{
the table without 0-sequences in var
}
\description{
Removes all rows with long 0-sequences in var from a table
}

