% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rm_nan}
\alias{rm_nan}
\title{Substitutes NaN in a vector with NA}
\usage{
rm_nan(x)
}
\arguments{
\item{x}{the vecor with NaN}
}
\value{
the vecor without NaN, but with NA instead
}
\description{
Substitutes NaN in a vector with NA
}

