% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_all}
\alias{run_all}
\title{Imputes the data in all csv-files of directory and writes new files}
\usage{
run_all(source_dir = "data/missing", target_dir = "data/imputed",
  vars = c("b", "g"), choices = get_std_choices(),
  pattern = "sensor-[0-9]*.csv", first_timestamp = "2014-01-01 00:00:00",
  last_timestamp = "2014-12-31 23:59:00", header = F, sep = ",",
  append = F)
}
\arguments{
\item{source_dir}{the directory of the input-csv-files}

\item{target_dir}{the directory where the result-csv-files should be written to}

\item{vars}{vector of the variablenames to be imputed}

\item{choices}{list of strategies, including gapsize-limits and function-parameters}

\item{pattern}{pattern of the input-file-names}

\item{first_timestamp}{first date of the result-data (as character)}

\item{last_timestamp}{last date of the result-data (as character)}

\item{header}{logical, if csv-input-file has header: T}

\item{sep}{the separator between columns in the csv-input}

\item{append}{logical, if all results should be written into one csv-file: T}
}
\value{
nothing, but writes new csv-files with imputed data to target_dir
}
\description{
Reads, preprocesses and imputes the data in all csv-files of a directory and writes new files to a subdirectory.
Default-strategies are those recommended in the text.
Only works for data-aggregation-interval of 5 minutes.
Lines in the input-csv files have to be in this format:
182,2014-01-01 00:01:00,0,0 (which means: sensor-id, timestamp, occupancy, speed).
Filenames of input-csv: sensor-x.csv (where x can be any number)
}

