% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{test_all}
\alias{test_all}
\title{Tests whole imputation-process and measures computing-time for different strategies}
\usage{
test_all(source_dir = "data/test", target_dir = "data/test_results",
  vars = c("b", "g"), choices = get_test_choices(),
  pattern = "sensor-[0-9]*.csv", first_timestamp = "2014-01-01 00:00:00",
  last_timestamp = "2014-12-31 23:59:00")
}
\arguments{
\item{source_dir}{the directory of the input-csv-files}

\item{target_dir}{the directory where the result-csv-files should be written to}

\item{vars}{vector of the variablenames to be imputed}

\item{choices}{list of lists of strategies (see get_test_choices())}

\item{pattern}{pattern of the input-file-names}

\item{first_timestamp}{first date of the result-data (as character)}

\item{last_timestamp}{last date of the result-data (as character)}
}
\value{
nothing, but writes new csv-files with imputed data to target_dir
}
\description{
Reads, preprocesses and imputes the data in all csv-files of a directory and writes new files to a subdirectory.
Measures computing-time for strategies and writes results to csv-file in target_dir.
Only works for data-aggregation-interval of 5 minutes.
Lines in the input-csv files have to be in this format:
182,2014-01-01 00:01:00,0,0 (which means: sensor-id, timestamp, occupancy, speed).
Filenames of input-csv: sensor-x.csv (where x can be any number).
}

